+++ extension/app/panel-popup/constants.js
@@ -2,7 +2,7 @@ var PANEL_POPUP_PAGE = "panel-popup-page.html";
 var PANEL_POPUP_ROOT_ID = "element-deleter-root";
 var PANEL_POPUP_HOST_ATTR = "data-element-deleter-ui";
 var PANEL_POPUP_SESSION_TAB_KEY = "panelPopupTab";
-var PANEL_POPUP_TABS = ["settings", "shortcuts", "info"];
+var PANEL_POPUP_TABS = ["settings", "shortcuts", "info", "recommend"];
 var PANEL_PAGE_CONFIG = {
   pageHtml: PANEL_POPUP_PAGE,
   sessionTabKey: PANEL_POPUP_SESSION_TAB_KEY,